pyproject.toml
[project]
name = "reddit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
    "rich>=13.9.4",
]
